+++ pyannote/metrics/errors/identification.py
@@ -3,7 +3,7 @@
 
 # The MIT License (MIT)
 
-# Copyright (c) 2012-2017 CNRS
+# Copyright (c) 2012-2019 CNRS
 
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -25,15 +25,13 @@
 
 # AUTHORS
 # Hervé BREDIN - http://herve.niderb.fr
-
-from __future__ import unicode_literals
+# Benjamin MAURICE - maurice@limsi.fr
 
 import numpy as np
-from munkres import Munkres
+from scipy.optimize import linear_sum_assignment
 
 from ..matcher import LabelMatcher
 from pyannote.core import Annotation
-from xarray import DataArray
 
 from ..matcher import MATCH_CORRECT, MATCH_CONFUSION, \
     MATCH_MISSED_DETECTION, MATCH_FALSE_ALARM
@@ -65,11 +63,9 @@ class IdentificationErrorAnalysis(UEMSupportMixin, object):
     def __init__(self, collar=0., skip_overlap=False):
 
         super(IdentificationErrorAnalysis, self).__init__()
-        self.matcher_ = LabelMatcher()
-        self.munkres = Munkres()
+        self.matcher = LabelMatcher()
         self.collar = collar
-        self.skip_overlap=skip_overlap
-
+        self.skip_overlap = skip_overlap
 
     def difference(self, reference, hypothesis, uem=None, uemified=False):
         """Get error analysis as `Annotation`
@@ -172,9 +168,7 @@ class IdentificationErrorAnalysis(UEMSupportMixin, object):
                 for i2, e2 in enumerate(new_errors):
                     match[i1, i2] = self._match_errors(e1, e2)
 
-            mapping = self.munkres.compute(2 - match)
-
-            for i1, i2 in mapping:
+            for i1, i2 in zip(*linear_sum_assignment(-match)):
 
                 if i1 >= n1:
                     track = behaviors.new_track(segment,
@@ -264,6 +258,16 @@ class IdentificationErrorAnalysis(UEMSupportMixin, object):
         ] + hLabels
 
         # initialize empty matrix
+
+        try:
+            from xarray import DataArray
+        except ImportError:
+            msg = (
+                "Please install xarray dependency to use class "
+                "'IdentificationErrorAnalysis'."
+            )
+            raise ImportError(msg)
+
         matrix = DataArray(
             np.zeros((len(rLabels), len(hLabels))),
             coords=[('reference', rLabels), ('hypothesis', hLabels)])